backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.4.0
Pillow==10.4.0
psycopg[binary]==3.2.1
python-dotenv==1.0.1
